pyproject.toml

[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.65.2",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=2.0.0",
    "requests>=2.32.3",
    "pytz>=2025.1",
    "joblib>=1.4.2",
    "tabulate>=0.9.0",
    "xgboost>=3.0.0",
    "flask-caching>=2.3.1",
    "sklearn-crfsuite>=0.5.0",
    "python-crfsuite>=0.9.11",
    "tqdm>=4.67.1",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.1",
    "tensorflow>=2.18.1",
    "numpy>=1.26.4",
    "scikit-learn==1.1.3",
    "scipy>=1.15.3",
    "pandas>=1.5.3",
]
